% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerQC.R
\name{check_maf}
\alias{check_maf}
\title{Identification of SNPs with a very low minor allele frequency}
\usage{
check_maf(qcdir, alg, mafTh = 0.01, macTh = 20, verbose = FALSE,
  interactive = FALSE, path2plink = NULL)
}
\arguments{
\item{qcdir}{[character] /path/to/directory/with/QC/results containing
alg.bim, alg.bed, alg.fam files and alg.fail.IDs containing IIDs of
individuals that failed QC. User needs writing permission to qcdir.}

\item{alg}{[character] prefix of QC-ed plink files, i.e. alg.bed, alg.bim,
alg.fam.}

\item{mafTh}{[double] Threshold for minor allele frequency cut-off.}

\item{macTh}{[double] Threshold for minor allele cut cut-off, if both mafTh
and macTh are specied, macTh is used (macTh = mafTh\*2\*NrSamples)}

\item{interactive}{[logical] Should plots be shown interactively? When
choosing this option, make sure you have X-forwarding/graphical interface
available for interactive plotting. Alternatively, set interactive=FALSE and
save the returned plot object (p_hwe) via ggplot2::ggsave(p=p_maf,
other_arguments) or pdf(outfile) print(p_maf) dev.off().}

\item{path2plink}{[character] Absolute path to where external plink software
\url{https://www.cog-genomics.org/plink/1.9/} can be found. If not provided,
assumed that PATH set-up works and plink will be found by system("plink").}
}
\value{
named list with i) fail_maf containing a [data.frame] with CHR
(Chromosome code), SNP (Variant identifier), A1 (Allele 1; usually minor), A2
(Allele 2; usually major), MAF (Allele 1 frequency), NCHROBS (Number of
allele observations) for all SNPs that failed the mafTh/macTh and ii) p_maf,
a ggplot2-object 'containing' the MAF distribution histogram which can be
shown by (print(p_maf)).
}
\description{
Uses plink --remove alg.fail.IDs --freq to calculate the minor allele
frequencies for all variants in the individuals that passed the perSampleQC.
The minor allele frequency distributions is plotted as a histogram.
}
