% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleQC.R
\name{check_relatedness}
\alias{check_relatedness}
\title{Identification of related individuals}
\usage{
check_relatedness(qcdir, alg, famfile = NULL, highIBDTh = 0.1875,
  interactive = FALSE)
}
\arguments{
\item{qcdir}{[character] path/to/directory/with/QC/results containing
alg.imiss and alg.genome results as returned by plink --missing and plink
--genome. Users needs writing permission to qcdir.}

\item{alg}{[character] prefix of QC-ed plink files, i.e. alg.bed, alg.bim,
alg.fam, alg.genome and alg.imiss.}

\item{famfile}{[character] path/to/file/with/additional/family/information.}

\item{highIBDTh}{[double] Threshold for acceptable proportion of IBD between
pair of individuals.}

\item{interactive}{[logical] Should plots be shown interactively? When
choosing this option, make sure you have X-forwarding/graphical interface
available for interactive plotting. Alternatively, set interactive=FALSE and
save the returned plot object (p_IBD() via ggplot2::ggsave(p=p_IBD,
other_arguments) or pdf(outfile) print(p_IBD) dev.off().}
}
\value{
a named [list] with i) fail_high_IBD containing a [data.frame] with
and ii) iii) p_IBD, a
ggplot2-object 'containing' all pair-wise IBD-estimates as histograms
stratified by value of PI_HAT, which can be
shown by print(p_IBD).
}
\description{
Evaluates and depicts results of plink --genome.
plink --genome calculates identity by state (IBS) for each pair of individuals
based on the average proportion of alleles shared at genotyped SNPs. The
degree of recent shared ancestry, i.e. the identity by descent (IBD) can
be estimated from the genome-wide IBS. The proportion of IBD between two
individuals is returned by --genome as PI_HAT.
check_relatedness uses two external scripts for parsing of the alg.genome
files to find pairs of samples whose proportion of IBD is larger than the
specified highIBDTh. Subsequently, the individual with the greater genotype
missingness rate is selected and returned as the individual failing the
relatedness check. Optionally, a famfile with additional
relatedness information can be provided.
check_relatedness depicts all pair-wise IBD-estimates as histograms
stratified by value of PI_HAT.
}
\examples{
package.dir <- find.package('plinkQC')
qcdir <- file.path(package.dir,'extdata')
alg <- 'data'
relatednessQC <- check_relatedness(qcdir=qcdir, alg=alg)
}
