% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerQC.R
\name{check_snp_missingness}
\alias{check_snp_missingness}
\title{Identification of SNPs with high missingness rates}
\usage{
check_snp_missingness(qcdir, alg, lmissTh = 0.01, interactive = FALSE,
  path2plink = NULL, verbose = FALSE, showPlinkError = TRUE)
}
\arguments{
\item{qcdir}{[character] /path/to/directory/with/QC/results containing
alg.bim, alg.bed, alg.fam files and alg.fail.IDs containing IIDs of
individuals that failed QC. User needs writing permission to qcdir.}

\item{alg}{[character] prefix of QC-ed plink files, i.e. alg.bed, alg.bim,
alg.fam.}

\item{lmissTh}{[double] Threshold for acceptable variant missing rate across
samples.}

\item{interactive}{[logical] Should plots be shown interactively? When
choosing this option, make sure you have X-forwarding/graphical interface
available for interactive plotting. Alternatively, set interactive=FALSE and
save the returned plot object (p_lmiss) via ggplot2::ggsave(p=p_lmiss,
other_arguments) or pdf(outfile) print(p_lmiss) dev.off().}

\item{path2plink}{[character] Absolute path to where external plink software
\url{https://www.cog-genomics.org/plink/1.9/} can be found. If not provided,
assumed that PATH set-up works and plink will be found by system("plink").}

\item{verbose}{[logical] If TRUE, progress info is printed to standard out
and specifically, if TRUE, plink log will be displayed.}
}
\value{
named list with i) fail_missingness containing a [data.frame] with
CHR (Chromosome code), SNP (Variant identifier), CLST (Cluster identifier.
Only present with --within/--family), N_MISS (Number of missing genotype
call(s), not counting obligatory missings), N_CLST  (Cluster size; does not
include nonmales on Ychr; Only present with --within/--family), N_GENO
(Number of potentially valid call(s)), F_MISS (Missing call rate) for all
SNPs failing the lmissTh and ii) p_lmiss, a ggplot2-object 'containing' the
SNP missingness histogram which can be shown by (print(p_lmiss)).
}
\description{
Uses plink --remove alg.fail.IDs --missing --freq to calculate the rates of
missing genotype calls and frequency for all variants in the individuals that
passed the perSampleQC. Depicts SNP missingness rates (stratified by
minor allele frequency) as histograms.
}
