% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry.R
\name{checkLoadingMat}
\alias{checkLoadingMat}
\title{Checking the path of the loading matrix}
\usage{
checkLoadingMat(path2load_mat)
}
\arguments{
\item{path2load_mat}{[character] /path/to/directory containing the filepath
to the loading matrices needed to run the ancestry functionality}
}
\description{
Makes sure that the loading matrix is located at the filepath stored in path2load_mat
}
