% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkPlink2}
\alias{checkPlink2}
\title{Check PLINK2 software access}
\usage{
checkPlink2(path2plink2 = NULL)
}
\arguments{
\item{path2plink2}{[character] Absolute path to PLINK executable
(\url{https://www.cog-genomics.org/plink/2.0/}) i.e.
plink 2 should be accessible as path2plink -h. The full name of the executable
should be specified: for windows OS, this means path/plink.exe, for unix
platforms this is path/plink. If not provided, assumed that PATH set-up works
and PLINK will be found by \code{\link[sys]{exec}}('plink').}
}
\value{
Path to PLINK version 2.0 executable.
}
\description{
checkPlink checks that the PLINK software version 2.0
(\url{https://www.cog-genomics.org/plink/2.0/})
can be found from system call.
}
