% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry.R
\name{checkRF_path}
\alias{checkRF_path}
\title{Checking the path of userinputted random forest}
\usage{
checkRF_path(rf_path, anc_list)
}
\arguments{
\item{rf_path}{[character] /path/to/directory where the rds file for a user inputted
classification model is located Note that the name of the file including the .rds or
.RDS extenstion must be included}

\item{anc_list}{[list] List of the names used for labels for the random forest}
}
\description{
Makes sure the user inputted random forest is able to be loaded
}
\examples{
rf_path <- '/path/to/model/model.rds'
\dontrun{
# the following code is not run on package build, as the path2load_mat on the
# user system is not known.
checkRF_path(rf_path = rf_path)
}
}
