% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleQC.R
\name{perSampleQC}
\alias{perSampleQC}
\title{Quality control for all individuals in plink-dataset}
\usage{
perSampleQC(qcdir, alg, do.check_sex = TRUE, maleTh = 0.8,
  femaleTh = 0.2, externalSex = NULL, externalMale = "M",
  externalSexSex = "Sex", externalSexID = "IID",
  externalFemale = "F", fixMixup = FALSE,
  do.check_heterozygosity_and_missingness = TRUE, imissTh = 0.03,
  hetTh = 3, do.check_relatedness = TRUE, highIBDTh = 0.1875,
  family = NULL, do.check_ancestry = TRUE, prefixMergedDataset,
  europeanTh = 1.5, refSamples = NULL, refColors = NULL,
  refSamplesFile = NULL, refColorsFile = NULL, refSamplesIID = "IID",
  refSamplesPop = "Pop", refColorsColor = "Color",
  refColorsPop = "Pop", studyColor = "#2c7bb6", interactive = FALSE,
  verbose = TRUE)
}
\arguments{
\item{qcdir}{[character] /path/to/directory/with/QC/results containing the
basic data files alg.bim, alg.bed, alg.fam files. In addition, if
do.[analysis] is set to TRUE, it expects the analysis-specific plink output
files in qcdir, i.e. do.check_sex expects alg.sexcheck,
do.check_heterozygosity_and_missingness expects alg.het and alg.imiss,
do.check_relatedness expects alg.genome and alg.imiss and do.check_ancestry
expects prefixMergeData.eigenvec.
User needs writing permission to qcdir.}

\item{alg}{[character] prefix of QC-ed plink files, i.e. alg.bed, alg.bim,
alg.fam.}

\item{do.check_sex}{[logical] if TRUE, run \code{\link{check_sex}}}

\item{maleTh}{[double] Threshold of X-chromosomal heterozygosity rate for
males.}

\item{femaleTh}{[double] Threshold of X-chromosomal heterozygosity rate for
females.}

\item{externalMale}{[integer/character] Identifier for 'male' in externalSex.}

\item{externalSexSex}{[character] Column identifier for column containing sex
information in externalSex.}

\item{externalSexID}{[character] Column identifier for column containing ID
information in externalSex.}

\item{externalFemale}{[integer/character] Identifier for 'female' in
externalSex.}

\item{fixMixup}{[logical] Should PEDSEX of individuals with mismatch between
PEDSEX and Sex, with Sex==SNPSEX automatically corrected: this will directly
change the alg.bim/.bed/.fam files!}

\item{do.check_heterozygosity_and_missingness}{[logical] if TRUE, run
\code{\link{check_heterozygosity_and_missingness}}.}

\item{imissTh}{[double] Threshold for acceptable missing genotype rate in any
individual}

\item{hetTh}{[double] Threshold for acceptable deviation from mean
heterozygosity in any individual. Expressed as multiples of standard
deviation of heterozygosity (het), i.e. individuals outside mean(het) +/-
hetTh*sd(het) will be returned as failing heterozygosity check.}

\item{do.check_relatedness}{[logical] if TRUE, run
\code{\link{check_relatedness}}.}

\item{highIBDTh}{[double] Threshold for acceptable proportion of IBD between
pair of individuals.}

\item{family}{[character] path/to/file/with/additional/family/information.}

\item{do.check_ancestry}{[logical] if TRUE, run
\code{\link{check_ancestry}}.}

\item{prefixMergedDataset}{[character] Prefix of merged dataset (study and
reference samples) used in plink --pca, resulting in
prefixMergedDataset.eigenvec}

\item{europeanTh}{[double] Scaling factor of radius to be drawn around center
of reference European samples, with study samples inside this radius
considered to be of European descent and samples outside this radius of
non-European descent. The radius is computed as the maximum Euclidean distance
of reference Europeans to the centre of reference European samples.}

\item{refSamplesFile}{[character] /Path/to/File/with/reference samples. Needs
columns with sample identifiers [refSamplesIID] corresponding to IIDs in
prefixMergedDataset.eigenvec and population identifier [refSamplesPop]
corresponding to population IDs [refColorsPop] in refColorsfile/refColors.}

\item{refColorsFile}{[character, optional]
/Path/to/File/with/Population/Colors; if not provided and is.null(refColors)
ggplot default colors for qualitative scale are used if plot is TRUE.}

\item{refSamplesIID}{[character] Column name of reference sample IDs in
refSamples/refSamplesFile.}

\item{refSamplesPop}{[character] Column name of reference sample population
IDs in refSamples/refSamplesFile.}

\item{refColorsColor}{[character] Column name of population colors in
refColors/refColorsFile}

\item{refColorsPop}{[character] Column name of reference sample population
IDs in refColors/refColorsFile.}

\item{studyColor}{[character] Color to be used for study population if plot
is TRUE.}

\item{interactive}{[logical] Should plots be shown interactively? When
choosing this option, make sure you have X-forwarding/graphical interface
available for interactive plotting. Alternatively, set interactive=FALSE and
save the returned plot object (p_sampleQC) via ggplot2::ggsave(p=p_sampleQC,
other_arguments) or pdf(outfile) print(p_sampleQC) dev.off().
If TRUE, i) depicts the X-chromosomal heterozygosity (SNPSEX) of the samples
split by their PEDSEX (if do.check_sex is TRUE), ii) creates a scatter plot
with samples' missingness rates on x-axis and their heterozygosity rates on
the y-axis (if do.check_heterozygosity_and_missingness is TRUE), iii) depicts
all pair-wise IBD-estimates as histogram (if do.check_relatedness is TRUE)
and iv) creates a scatter plot of PC1 versus PC2 color-coded for samples of
reference populations and study population (if do.check_relatedness is set to
TRUE).}

\item{path2plink}{[character] Absolute path to where external plink software
\url{https://www.cog-genomics.org/plink/1.9/} can be found. If not provided,
assumed that PATH set-up works and plink will be found by system("plink").}
}
\value{
named [list] with i) sample_missingness containing a [vector] with
sample IIDs failing the missingness threshold imissTh, ii) highIBD containing
a [vector] with sample IIDs failing the relatedness threshold highIBDTh, iii)
outlying_heterozygosity containing a [vector] with sample IIDs failing the
heterozygosity threshold hetTh, iv) mismatched_sex containing a [vector] with
the sample IIDs failing the sexcheck based on SNPSEX and femaleTh/maleTh and
v) ancestry containing a vector with sample IIDs failing the ancestry check
based on europeanTh and vi) p_sampleQC, a ggplot2-object 'containing' a
sub-paneled plot with the QC-plots of \code{\link{check_sex}},
\code{\link{check_heterozygosity_and_missingness}},
\code{\link{check_relatedness}} and \code{\link{check_ancestry}}, which can
be shown by print(p_sampleQC).
List entries contain NULL if that specific check was not chosen.
}
\description{
perSampleQC checks the samples in the plink dataset for their total
missingness and heterozygosity rates, the concordance of their assigned sex
to their SNP sex and their genetic ancestry.
}
\details{
perSampleQC wraps around the individual QC functions
\code{\link{check_sex}}, \code{\link{check_heterozygosity_and_missingness}},
\code{\link{check_relatedness}} and \code{\link{check_ancestry}}. For details
on the parameters and outputs, check these function documentations. For
detailed output for fail IIDs (instead of simple IID lists), run each
function individually.
}
