% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkQC.R
\docType{package}
\name{plinkQC}
\alias{plinkQC-package}
\alias{plinkQC}
\title{plinkQC}
\description{
Genotype Quality Control with 'PLINK'
}
\seealso{
Useful links:
\itemize{
  \item \url{https://meyer-lab-cshl.github.io/plinkQC/}
  \item Report bugs at \url{https://github.com/meyer-lab-cshl/plinkQC/issues}
}

}
\author{
\strong{Maintainer}: Hannah Meyer \email{hannah.v.meyer@gmail.com} (\href{https://orcid.org/0000-0003-4564-0899}{ORCID})

Other contributors:
\itemize{
  \item Caroline Walter \email{carowalter26@hotmail.com} [contributor]
}

}
\keyword{internal}
