% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry.R
\name{run_ancestry_format}
\alias{run_ancestry_format}
\title{Running functions to format data for ancestry identification}
\usage{
run_ancestry_format(
  indir,
  name,
  qcdir = indir,
  verbose = FALSE,
  path2plink2 = NULL,
  keep_individuals = NULL,
  remove_individuals = NULL,
  exclude_markers = NULL,
  extract_markers = NULL,
  showPlinkOutput = TRUE,
  format = "@:#[hg38]",
  plink2format = FALSE,
  var_format = FALSE,
  path2load_mat,
  legend_text_size,
  legend_title_size,
  axis_text_size,
  axis_title_size,
  title_size,
  legend_position = "bottom"
)
}
\arguments{
\item{indir}{[character] /path/to/directory containing the basic PLINK 1.9 data
file name.bim, name.fam, name.bed}

\item{name}{[character] Prefix of PLINK 1.9 files, i.e. name.bim, name.fam, 
name.bed}

\item{qcdir}{[character] /path/to/directory where the plink2 data formations
as returned by plink2 --make-pgen will be saved to. User needs writing 
permission to qcdir. Per default is qcdir=indir.}

\item{verbose}{[logical] If TRUE, progress info is printed to standard out.}

\item{path2plink2}{[character] Absolute path to PLINK executable
(\url{https://www.cog-genomics.org/plink/2.0/}) i.e.
plink 2 should be accessible as path2plink -h. The full name of the executable
should be specified: for windows OS, this means path/plink.exe, for unix
platforms this is path/plink. If not provided, assumed that PATH set-up works
and PLINK will be found by \code{\link[sys]{exec}}('plink').}

\item{keep_individuals}{[character] Path to file with individuals to be
retained in the analysis. The file has to be a space/tab-delimited text file
with family IDs in the first column and within-family IDs in the second
column. All samples not listed in this file will be removed from the current
analysis. See \url{https://www.cog-genomics.org/plink/1.9/filter#indiv}.
Default: NULL, i.e. no filtering on individuals.}

\item{remove_individuals}{[character] Path to file with individuals to be
removed from the analysis. The file has to be a space/tab-delimited text file
with family IDs in the first column and within-family IDs in the second
column. All samples listed in this file will be removed from the current
analysis. See \url{https://www.cog-genomics.org/plink/1.9/filter#indiv}.
Default: NULL, i.e. no filtering on individuals.}

\item{exclude_markers}{[character] Path to file with makers to be
removed from the analysis. The file has to be a text file with a list of
variant IDs (usually one per line, but it's okay for them to just be
separated by spaces). All listed variants will be removed from the current
analysis. See \url{https://www.cog-genomics.org/plink/1.9/filter#snp}.
Default: NULL, i.e. no filtering on markers.}

\item{extract_markers}{[character] Path to file with makers to be
included in the analysis. The file has to be a text file with a list of
variant IDs (usually one per line, but it's okay for them to just be
separated by spaces). All unlisted variants will be removed from the current
analysis. See \url{https://www.cog-genomics.org/plink/1.9/filter#snp}.
Default: NULL, i.e. no filtering on markers.}

\item{showPlinkOutput}{[logical] If TRUE, plink log and error messages are
printed to standard out.}

\item{format}{[character] This gives the template to rewrite the variant identifier.
A '@' represents the chromosome code, and a '#' represents the base-pair position.}

\item{plink2format}{[logical] If TRUE, data is in plink2 format already and 
convert_to_plink2 will not be run}

\item{var_format}{[logical] If TRUE, variant identifiers are in correct 
format already and rename_variant_identifiers will not be run}

\item{path2load_mat}{[character] /path/to/directory where loading matrices are 
kept. This can be downloaded from: https://github.com/meyer-lab-cshl/plinkQCAncestryData}

\item{legend_text_size}{[integer] Size for legend text.}

\item{legend_title_size}{[integer] Size for legend title.}

\item{axis_text_size}{[integer] Size for axis text.}

\item{axis_title_size}{[integer] Size for axis title.}

\item{title_size}{[integer] Size for plot title.}

\item{legend_position}{[character] Legend position for the plot.}
}
\value{
Creates plink 2.0 datafiles
}
\description{
This function runs convert_to_plink2 and rename_variant_identifiers to format 
the data for the ancestry identification with superpop_classification
}
\examples{
indir <- system.file("extdata", package="plinkQC")
qcdir <- tempdir()
name <- "data"
path2plink <- '/path/to/plink'
# the following code is not run on package build, as the path2plink on the
# user system is not known.
\dontrun{
run_ancestry_format(indir=indir, qcdir=qcdir, 
  name=name, path2plink2 = path2plink2)
}
}
