% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualQC.R
\name{run_check_ancestry}
\alias{run_check_ancestry}
\title{Run PLINK principal component analysis}
\usage{
run_check_ancestry(qcdir, prefixMergedDataset, verbose = FALSE,
  path2plink = NULL, showPlinkOutput = TRUE)
}
\arguments{
\item{qcdir}{[character] /path/to/directory/with/QC/results containing the
basic data files alg.bim, alg.bed, alg.fam files. User needs writing
permission to qcdir.}

\item{prefixMergedDataset}{[character] Prefix of merged study and
reference data files, i.e. prefixMergedDataset.bed, prefixMergedDataset.bim,
prefixMergedDataset.fam.}

\item{verbose}{[logical] If TRUE, progress info is printed to standard out.}

\item{path2plink}{[character] Absolute path to directory where external plink
software \url{https://www.cog-genomics.org/plink/1.9/} can be found, i.e.
plink should be accesible as path2plink/plink -h. If not
provided, assumed that PATH set-up works and plink will be found by
system("plink").}

\item{showPlinkOutput}{[logical] If TRUE, plink log and error messages are
printed to standard out.}
}
\description{
Run plink --pca to calculate the principal components on merged genotypes
of the study and reference dataset.
}
\details{
Both, \code{\link{run_check_ancestry}} and its evaluation by
\code{\link{evaluate_check_ancestry}} can simply be invoked by
\code{\link{check_ancestry}}.
}
\examples{
package.dir <- find.package('plinkQC')
qcdir <- file.path(package.dir,'extdata')
prefixMergedDataset <- 'data.HapMapIII'
# the following code is not run on package build, as the path2plink on the
# user system is not known.
\dontrun{
run <- run_check_ancestry(qcdir, prefixMergedDataset)
}
}
