% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualQC.R
\name{run_check_sex}
\alias{run_check_sex}
\title{Run PLINK sexcheck.}
\usage{
run_check_sex(qcdir, alg, verbose = FALSE, path2plink = NULL,
  showPlinkOutput = TRUE)
}
\arguments{
\item{qcdir}{[character] /path/to/directory/with/QC/results containing the
basic data files alg.bim, alg.bed, alg.fam files. User needs writing
permission to qcdir.}

\item{alg}{[character] prefix of QC-ed plink files, i.e. alg.bed, alg.bim,
alg.fam.}

\item{verbose}{[logical] If TRUE, progress info is printed to standard out.}

\item{path2plink}{[character] Absolute path to directory where external plink
software \url{https://www.cog-genomics.org/plink/1.9/} can be found, i.e.
plink should be accesible as path2plink/plink -h. If not
provided, assumed that PATH set-up works and plink will be found by
system("plink").}

\item{showPlinkOutput}{[logical] If TRUE, plink log and error messages are
printed to standard out.}
}
\description{
Run  plink --sexcheck to calculate the heterozygosity rate
across X-chromosomal variants.
}
\details{
Both \code{\link{run_check_sex}} and its evaluation
\code{\link{evaluate_check_sex}} can simply be invoked by
\code{\link{check_sex}}).
}
\examples{
package.dir <- find.package('plinkQC')
qcdir <- file.path(package.dir,'extdata')
alg <- 'data'
# the following code is not run on package build, as the path2plink on the
# user system is not known.
\dontrun{
run <- run_check_heterozygosity(qcdir, alg)
}
}
