% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testNumerics}
\alias{testNumerics}
\title{Test lists for different properties of numerics}
\usage{
testNumerics(numbers, positives = NULL, integers = NULL,
  proportions = NULL)
}
\arguments{
\item{numbers}{[list] whose elements are tested for being numeric.}

\item{positives}{[list] whose elements are tested for being positive numbers.}

\item{integers}{[list] whose elements are tested for being integers.}

\item{proportions}{[list] whose elements are tested for being proportions.
between 0 and 1.}
}
\description{
Test all elements of a list if they are numeric, positive numbers, integers
or proportions (range 0-1).
}
