% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry.R
\name{checkLoadingMat}
\alias{checkLoadingMat}
\title{Checking the path of the loading matrix}
\usage{
checkLoadingMat(path2load_mat)
}
\arguments{
\item{path2load_mat}{[character] /path/to/directory where loading matrices are 
kept. This can be downloaded from the github repo. Note that the name of the file 
before the .eigenvec.allele or .acount must be included in file path.}
}
\description{
Makes sure that the loading matrix is located at the filepath stored in path2load_mat
}
