% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkPlink}
\alias{checkPlink}
\title{Check PLINK software access}
\usage{
checkPlink(path2plink = NULL)
}
\arguments{
\item{path2plink}{[character] Absolute path to PLINK executable
(\url{https://www.cog-genomics.org/plink/1.9/}) i.e.
plink should be accesible as path2plink -h. The full name of the executable
has to be specified: for windows OS, this means path/plink.exe, for unix
platforms this is path/plink. If not provided, assumed that PATH set-up works
and PLINK will be found by \code{\link[sys]{exec_wait}('plink')}.}
}
\value{
if PLINK can be found and excuted, returns
'PLINK can be found and executed.' else fails with appropriate error message.
}
\description{
checkPlink checks that the PLINK software
(\url{https://www.cog-genomics.org/plink/1.9/})
can be found from system call.
}
