% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkPlink}
\alias{checkPlink}
\title{Check PLINK software access}
\usage{
checkPlink(path2plink)
}
\arguments{
\item{path2plink}{[character, optional] Absolute path to plink software. If
not provided, assumed that PATH set-up works and plink will be found by
system("plink").}
}
\description{
checkPlink checks that the PLINK software
(\url{https://www.cog-genomics.org/plink/1.9/}) required for perMarkerQC and
check_sex can be found from system call
}
