% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyQC.R
\name{clean_data}
\alias{clean_data}
\title{Create plink dataset with samples and markers passing quality control}
\usage{
clean_data(qcdir, alg, macTh = 20, mafTh = 0.01, highIBDTh = NULL)
}
\description{
Remove samples that fail per individual QC
Remove markers that fail per marker QC
}
