% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerQC.R
\name{perMarkerQC}
\alias{perMarkerQC}
\title{Quality control for all markers in plink-dataset}
\usage{
perMarkerQC(qcdir, alg, mafTh = 0.01, macTh = 20, hweTh = 1e-05,
  lmissTh = 0.01, interactive = FALSE, verbose = TRUE,
  path2plink = NULL, showPlinkOutput = TRUE)
}
\arguments{
\item{qcdir}{[character] /path/to/directory/with/QC/results containing
alg.bim, alg.bed, alg.fam files and alg.fail.IDs containing IIDs of
individuals that failed QC. User needs writing permission to qcdir.}

\item{alg}{[character] prefix of QC-ed plink files, i.e. alg.bed, alg.bim,
alg.fam.}

\item{mafTh}{[double] Threshold for minor allele frequency cut-off.}

\item{macTh}{[double] Threshold for minor allele cut cut-off, if both mafTh
and macTh are specified, macTh is used (macTh = mafTh\*2\*NrSamples)}

\item{hweTh}{[double] Significance threshold for deviation from HWE.}

\item{lmissTh}{[double] Threshold for acceptable variant missing rate across
samples.}

\item{interactive}{[logical] Should plots be shown interactively? When
choosing this option, make sure you have X-forwarding/graphical interface
available for interactive plotting. Alternatively, set interactive=FALSE and
save the returned plot object (p_marker) via ggplot2::ggsave(p=p_marker,
other_arguments) or pdf(outfile) print(p_marker) dev.off().}

\item{verbose}{[logical] If TRUE, progress info is printed to standard out.}

\item{path2plink}{[character] Absolute path to where external plink software
\url{https://www.cog-genomics.org/plink/1.9/} can be found. If not provided,
assumed that PATH set-up works and plink will be found by system("plink").}

\item{showPlinkOutput}{[logical] If TRUE, plink log and error messages are
printed to standard out.}
}
\value{
named [list] with i) fail_list, a named [list] with 1.
SNP_missingness, containing SNP IDs failing the missingness threshold
lmissTh, 2. hwe, containing SNP IDs failing the HWE exact test threshold
hweTh and 3. maf, containing SNPs failing the MAF threshold mafTh/MAC
threshold macTh and ii) p_markerQC, a ggplot2-object 'containing' a
sub-paneled plot with the QC-plots of \code{\link{check_snp_missingness}},
\code{\link{check_hwe}} and \code{\link{check_maf}}, which can
be shown by print(p_markerQC).
List entries contain NULL if that specific check was not chosen.
}
\description{
perMarkerQC checks the markers in the plink dataset for their missingness
rates across samples, their deviation from Hardy-Weinberg-Equilibrium (HWE)
and their minor allele frequencies (MAF). Per default, it assumes that IDs of
individuals that have failed \code{\link{perSampleQC}} have been written
to qcdir/alg.failIDs and removes these individuals when computing missingness
rates, HWE p-values and MAF. If the qcdir/alg.failIDs file does not exist, a
message is written to stdout but the analyses will continue for all samples
in the alg.fam/al.bed/alg.bim dataset.
Depicts i) SNP missingness rates (stratified by minor allele
frequency) as histograms, ii) p-values of HWE exact test (stratified by all
and low p-values) as histograms and iii) the minor allele frequency
distribution as a histogram.
}
\details{
perMarkerQC wraps around the marker QC functions
\code{\link{check_snp_missingness}}, \code{\link{check_hwe}} and
\code{\link{check_maf}}. For details on the parameters and outputs, check
these function documentations.
}
